// src/services/types/Genre.ts
enum Genre {
    War = "War",
    Satire = "Satire",
    Comedy = "Comedy",
    Novel = "Novel",
    Modernist = "Modernist",
    Tragedy = "Tragedy",
    GreatAmericanNovel = "Great American Novel",
    Historical = "Historical",
    Realistic = "Realistic",
    ComingOfAge = "Coming of Age",
    Romance = "Romance",
    SocialCommentary = "Social Commentary",
    Dystopian = "Dystopian",
    AfricanAmerican = "African American",
    Absurdist = "Absurdist",
    Philosophical = "Philosophical",
}

export default Genre;